+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed route from a store to a state
+- Changed link from an Action to an Attachment
+- Changed link from using addEventListener to on (svelte/events)
 
 ### Removed